Refactor library view into shared module
