+++ assets/js/app-shell.js
@@ -1,54 +1,15 @@
 import { initNavigation, loadToy, loadFromQuery } from './loader.js';
+import { createLibraryView } from './library-view.js';
 import toysData from './toys-data.js';
 
-let allToys = [];
+const libraryView = createLibraryView({
+  toys: toysData,
+  loadToy,
+  initNavigation,
+  loadFromQuery,
+  targetId: 'toy-list',
+  searchInputId: 'search-bar',
+  cardElement: 'div',
+});
 
-function createCard(toy) {
-  const card = document.createElement('div');
-  card.className = 'webtoy-card';
-  const title = document.createElement('h3');
-  title.textContent = toy.title;
-  const desc = document.createElement('p');
-  desc.textContent = toy.description;
-  card.appendChild(title);
-  card.appendChild(desc);
-
-  card.addEventListener('click', () => {
-    if (toy.type === 'module') {
-      loadToy(toy.slug, { pushState: true });
-    } else {
-      window.location.href = toy.module;
-    }
-  });
-
-  return card;
-}
-
-function renderToys(toys) {
-  const list = document.getElementById('toy-list');
-  if (!list) return;
-  list.innerHTML = '';
-  toys.forEach((toy) => list.appendChild(createCard(toy)));
-}
-
-function filterToys(query) {
-  const search = query.toLowerCase();
-  const filtered = allToys.filter(
-    (t) =>
-      t.title.toLowerCase().includes(search) ||
-      t.description.toLowerCase().includes(search)
-  );
-  renderToys(filtered);
-}
-
-async function init() {
-  allToys = toysData;
-  renderToys(allToys);
-
-  const search = document.getElementById('search-bar');
-  search?.addEventListener('input', (e) => filterToys(e.target.value));
-  initNavigation();
-  await loadFromQuery();
-}
-
-init();
+libraryView.init();
